requirements.txt
wheel
MySQL-python==1.2.5
pip-tools==0.3.4
django==1.6.5
psycopg2==2.5.3
South==1.0
Unipath==1.0
Pillow==2.4.0
django-crispy-forms==1.4.0
pyOpenSSL==0.14 # for requests to support SNI
ndg-httpsclient==0.3.2 # for requests to support SNI
pyasn1==0.1.7 # for requests to support SNI
requests==2.3.0
django-mptt==0.6.1
celery==3.1.12
django-celery==3.1.10
django-countries==2.1.2
django_tables2==0.15.0
django-tables2-reports==0.0.9
django-extra-views==0.6.4
Django-Select2==4.2.2
django-sitetree==1.0.0
django-admin-honeypot==0.3.0
django-rosetta==0.7.4
fabric==1.9.0
django-braces==1.4.0
python-social-auth==0.1.26
django-jsonfield==0.9.12
django-grappelli==2.5.3
easy-thumbnails==2.0.1
ipython==2.1.0
xlwt==0.7.5
reportlab==3.1.8
python-memcached==1.53
django-save-the-change==1.0.0
pwgen==0.4
premailer==2.5.0
django-redactoreditor==1.2.7
django-auth-policy==0.9.5
django-impersonate==0.8.1
PyYAML==3.11
django-mailgun==0.2.2
django-ssl-slapper==1.1
pyinotify==0.9.4
django-sekizai==0.7
django-extensions
django-debug-toolbar
django-debug-panel